♻️ reorganize browser-native modules by SDK domain
