+++ packages/browser-native/src/page.ts
@@ -1,0 +1,9 @@
+/** Returns document.readyState. */
+export function getReadyState(): DocumentReadyState {
+  return document.readyState
+}
+
+/** Returns document.visibilityState. */
+export function getVisibilityState(): DocumentVisibilityState {
+  return document.visibilityState
+}
