+++ packages/browser-native/src/timing.ts
